django/requirements.txt
asgiref==3.3.4
Django==3.0.8
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
Markdown==3.3.4
psycopg2==2.8.5
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
unicodecsv==0.14.1
